fix: client side auth
